Change: (burly-url-buffer) Workaround for unfound buffer names
If a buffer can't be found by name, rather than returning nil (which
bubbles up and causes an error when Emacs tries to put the window
state), make a buffer explaining the problem and return it.

This kind of error usually represents a bug in the mode of the buffer
that was bookmarked (i.e. a bad or missing
bookmark-make-record-function) rather than in Burly itself, and we
don't want that to abort the opening of other buffers/windows for the
bookmark.